libcpu: Cast struct timespec fields to uint64_t to avoid overflow.

+++ libcpu/timings.cpp
@@ -15,7 +15,7 @@ uint64_t abs_time() {
   struct timespec ts;
   if (clock_gettime(CLOCK_MONOTONIC, &ts) < 0)
     return 0;
-  return ts.tv_sec * 1000000000 + ts.tv_nsec;
+  return (uint64_t)ts.tv_sec * 1000000000 + (uint64_t)ts.tv_nsec;
 }
 
 #endif
